only try to fix if adding, not updating

+++ fixSystemErrors.js
@@ -12,22 +12,26 @@ const JobError = require('./models/JobError');
     let deletes = [];
 
     errors.forEach(async error => {
-        let split = error.message.split('Error adding item ');
-        let secondSplit = split[1];
-        let thirdSplit = secondSplit.split(', ');
-        let itemId = thirdSplit[0];
-        let fourthSplit = thirdSplit[1].split(' - ');
-        let listingId = fourthSplit[0];
+        if (error.message.indexOf('Error adding item') !== -1) {
+            let split = error.message.split('Error adding item ');
+            let secondSplit = split[1];
+            let thirdSplit = secondSplit.split(', ');
+            let itemId = thirdSplit[0];
+            let fourthSplit = thirdSplit[1].split(' - ');
+            let listingId = fourthSplit[0];
 
-        console.log(error._id);
+            console.log(error._id);
 
-        jobs.push({ type: 'add', itemId: itemId, listingId: listingId });
-        deletes.push(JobError.deleteOne({ _id: error._id }));
+            jobs.push({type: 'add', itemId: itemId, listingId: listingId});
+            deletes.push(JobError.deleteOne({_id: error._id}));
+        }
 });
 
     console.log(jobs);
 
     queue.itemQueue.add({ jobs: jobs });
 
-    await Promise.all(deletes);
+    if (deletes.length) {
+        await Promise.all(deletes);
+    }
 })();
